requirements.txt
pyyaml
paho-mqtt
pyserial